Clean up dead approaches and combine WS and handlers
Having a single app will be easier to manage, and we only
need a thin WS app if we hit scale issues. Things are written
to be pretty easy to swap out and return to the old setup.
