+++ .gitignore
@@ -1,2 +1,5 @@
 .env
 node_modules
+dist
+build
+.idea
