--- resistbot-flask/README.md
@@ -1,14 +1,0 @@
-## Overview
-This is a small flask app that can handle:
-
-1. Getting DMs from a discord user and forwarding them to RapidPro
-2. Forwarding a message from RapidPro to a discord user as a DM
-3. Receiving quick replies and forwarding them to RapidPro
-
-The idea was that each function could be a lambda, though I haven't used
-lambdas before so I'm not sure I mapped things correctly.
-
-Still lots of work to make this production ready.
-
-## How to run this
-See the root README
