--- resistbot-gateway-forwarder/.gitignore
@@ -1,3 +1,0 @@
-build
-node_modules
-.idea
