--- resistbot-gateway-forwarder/README.md
@@ -1,4 +1,0 @@
-## Overview
-This is a tiny node server that establishes a websocket connection
-and forwards DMs and [interactions](https://discord.com/developers/docs/interactions/message-components)
-to the flask application.
