--- resistbot-gateway-forwarder/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "resistbot-gateway-forwarder",
-  "version": "0.1.0",
-  "description": "Maintains a WS connection with discord and forwards relevant messages",
-  "main": "build/index.js",
-  "scripts": {
-    "build": "tsc",
-    "start": "node build/index.js",
-    "dev": "tsc && dotenv -e ../.env node build/index.js",
-    "docker-build": "docker build . -t resistbot-gateway-forwarder",
-    "docker-run": "docker run -d --env-file ../.env -t resistbot-gateway-forwarder -d | xargs docker logs -f",
-    "docker-stop-all": "docker ps --format json | jq '.ID' | xargs docker stop",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "discord-api-types": "^0.37.67",
-    "discord.js": "^14.14.1",
-    "typescript": "^5.3.3"
-  },
-  "devDependencies": {
-    "dotenv-cli": "^7.3.0",
-    "prettier": "^3.1.1",
-    "typescript": "^5.3.3"
-  }
-}
