--- resistbot-node-backend/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "resistbot-node-backend",
-  "main": "dist/index.js",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "build": "npx tsc",
-    "start": "node dist/index.js",
-    "dev": "dotenv -e ../.env nodemon src/index.ts",
-    "docker-build": "docker build . -t resistbot-node-backend",
-    "//docker-run": "Port 5555 is hard coded here and is the default. Reading it from .env seems like overkill.",
-    "docker-run": "docker run -d --env-file ../.env -p 5555:5555 -t resistbot-node-backend | xargs docker logs -f",
-    "docker-stop-all": "docker ps --format json | jq '.ID' | xargs docker stop",
-    "test": "vitest --config ./vitest/vi.config.ts"
-  },
-  "dependencies": {
-    "cookie-parser": "~1.4.4",
-    "debug": "~2.6.9",
-    "express": "^4.18.2",
-    "morgan": "~1.9.1"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.21",
-    "@types/morgan": "^1.9.9",
-    "@types/node": "^20.11.19",
-    "dotenv-cli": "^7.3.0",
-    "nodemon": "^3.0.3",
-    "prettier": "3.2.5",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.3",
-    "vitest": "^1.4.0",
-    "vitest-fetch-mock": "^0.2.2"
-  }
-}
