+++ src/utils/discord.ts
@@ -1,0 +1,7 @@
+export function createEmbed(url: string) {
+  return { url: url, title: url };
+}
+
+export function createButtonComponent(buttonText: string) {
+  return { type: 2, style: 1, label: buttonText, custom_id: buttonText };
+}
