+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "es2016",
+    "target": "es2017",
     "module": "commonjs",
     "outDir": "./dist",
     "esModuleInterop": true,
@@ -8,5 +8,7 @@
     "strict": true,
     "noImplicitAny": true,
     "skipLibCheck": true
-  }
+  },
+  "exclude": ["node_modules", "build", "test"],
+  "include": ["src"]
 }
