+++ types/global.d.ts
@@ -1,2 +1,2 @@
 // This is supposed to make typescript happy with test files but it doesn't seem to be!
-import 'vitest-fetch-mock';
+import "vitest-fetch-mock";
