+++ vitest/setupVitest.js
@@ -1,5 +1,5 @@
-import createFetchMock from 'vitest-fetch-mock';
-import { vi } from 'vitest';
+import createFetchMock from "vitest-fetch-mock";
+import { vi } from "vitest";
 
 const fetchMocker = createFetchMock(vi);
